requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
PyMySQL==1.1.1
typing_extensions==4.12.2
Werkzeug==3.0.4
mysql-connector-python
SQLAlchemy==1.4.31

